requirements.txt
flask==3.1.0
flask-cors==5.0.1
python-dotenv==1.1.0
langchain==0.3.22
langgraph==0.3.22
chromadb==0.6.3
sentence-transformers==2.5.1
openai==1.70.0
google-generativeai==0.4.0
pillow==10.2.0
python-multipart==0.0.9
pydantic==2.11.1
numpy==2.0.2
pandas==2.2.3
unstructured==0.12.5
pdf2image==1.17.0
pytesseract==0.3.10
python-magic==0.4.27
werkzeug==3.1.3
PyPDF2==3.0.1
pyyaml==6.0.1
python-docx==1.1.0
openpyxl==3.1.2
markdown==3.5 